fix: Make estimate_arrow_bytes_for_file infallible

+++ compactor/src/cold.rs
@@ -331,15 +331,15 @@ mod tests {
         table.create_column("time", ColumnType::Time).await;
         let table_column_types = vec![
             ColumnTypeCount {
-                col_type: ColumnType::Tag as i16,
+                col_type: ColumnType::Tag,
                 count: 3,
             },
             ColumnTypeCount {
-                col_type: ColumnType::I64 as i16,
+                col_type: ColumnType::I64,
                 count: 1,
             },
             ColumnTypeCount {
-                col_type: ColumnType::Time as i16,
+                col_type: ColumnType::Time,
                 count: 1,
             },
         ];
@@ -572,15 +572,15 @@ mod tests {
         table.create_column("time", ColumnType::Time).await;
         let table_column_types = vec![
             ColumnTypeCount {
-                col_type: ColumnType::Tag as i16,
+                col_type: ColumnType::Tag,
                 count: 3,
             },
             ColumnTypeCount {
-                col_type: ColumnType::I64 as i16,
+                col_type: ColumnType::I64,
                 count: 1,
             },
             ColumnTypeCount {
-                col_type: ColumnType::Time as i16,
+                col_type: ColumnType::Time,
                 count: 1,
             },
         ];
